requirements.txt
asgiref==3.11.1
cloudinary==1.44.1
Django==6.0.3
sqlparse==0.5.5
tzdata==2025.3
django-auditlog==3.4.1
django-autocomplete-light==3.12.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
django-filter==25.2
django-nested-admin==4.1.6
djangorestframework==3.16.1
djangorestframework-gis==1.2.0
djangorestframework_simplejwt==5.5.1
python-decouple==3.8
redis==5.3.0
requests==2.31.0
whitenoise==6.11.0
daphne==4.2.1
channels-redis==4.0.0
dj-database-url==3.1.0
psycopg2-binary==2.9.11
resend==2.23.0
